# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlow API
  description: Secure, self-hosted family data vault API
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:3460
    description: Development server
paths:
...
  /api/items:
    get:
      summary: List items
      tags: [Items]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of items
    post:
      summary: Create an item
      tags: [Items]
      responses:
        '201':
          description: Item created
...
  /api/items/{id}:
    get:
      summary: Get an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item details
    put:
      summary: Update an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item updated
    delete:
      summary: Delete an item (soft delete)
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item deleted
  /api/items/{id}/restore:
    post:
      summary: Restore item from trash
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item restored
  /api/items/{id}/favorite:
    post:
      summary: Toggle favorite
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Favorite toggled
    put:
      summary: Set favorite status
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Favorite set
  /api/items/{id}/copy:
    post:
      summary: Copy an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Item copied
  /api/items/{id}/history:
    get:
      summary: Get item history
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item history
  /api/items/{id}/share:
    post:
      summary: Share an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item shared
  /api/items/{id}/share/{shareUserId}:
    delete:
      summary: Unshare an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: shareUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item unshared
  /api/items/{id}/shares:
    get:
      summary: List item shares
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of shares
...